refactor(v4): file the last copies beside the family that uses them
`migration/utils/` held what several families were expected to share.
Three copies are left and each has exactly one caller, so the directory
is gone:

- `keyframes.ts` and the `cubicBezier` beside it → `ScrollAnimation/`,
  the only family that samples a keyframe. Motion's scroll-linked
  animation makes that family obsolete, and the interpolator now leaves
  in the same deletion instead of outliving its consumer in a shared
  folder.
- `deepmerge.ts` → `Track/payload.ts`. Merging analytics layers is a
  Track concern spelled as a utility: `AbstractTrack` and `TrackContext`
  are its only callers, and no native merge replaces it — a spread is
  shallow, so an ancestor's `ecommerce.currency` would vanish under a
  child's `ecommerce.items`.

Each family barrel takes over what the migration barrel re-exported, and
the barrel stops re-exporting `src/utils/maths.js`, which it only ever
did because the maths were a copy too.
