+++ packages/v4/DESIGN.md
@@ -849,7 +849,7 @@ This is already the de facto state: `src/` contains no animation utility of any 
 
 **The keyframes interpolator is not promoted (2026-08-16), because what needs it is going away.** It was on the list above — `compile(keyframes, { easing }) => (progress, size) => styles`, ~150 lines including the `cubicBezier` which replaces `@motionone/easing`. `ui-motion` does not ship an interpolator either: it carries Motion's **scroll-linked animation**, which is the whole job the `ScrollAnimation` family exists to do. Those components become obsolete, and the interpolator with them — so promoting it into core would be hoisting a primitive whose only consumer is scheduled for deletion.
 
-Core keeps what the rest of the ported components need from the maths: the easing functions on `utils/easings.ts` (ported 2026-08-16), `spring()` and `smoothTo()`. `migration/utils/keyframes.ts` and the `cubicBezier` beside it stay in `migration/` until the `ScrollAnimation` family goes.
+Core keeps what the rest of the ported components need from the maths: the easing functions on `utils/easings.ts` (ported 2026-08-16), `spring()` and `smoothTo()`. The interpolator and its bezier live in `migration/ScrollAnimation/`, beside the only family that calls them, so they leave with it in one deletion.
 
 **Out of core, into a separate `ui-animation` package:** time-based playback, stagger, sequencing, morphing and text splitting. Springs were on this list and came off it (2026-08-12): `spring()` is forty lines of pure maths with no player, no registry and no scheduling of its own, and `smoothTo()` is the one primitive the ported components need to smooth a value towards a target. What belongs outside is the _engine_ — a timeline, playback controls, a per-element registry — not a function that advances one number by one step. Two entry points over one package — Motion as declarative components (`data-component="Motion"`, its own props as `data-option-*`), GSAP as a lifecycle/scoping decorator (`gsap.context()` bound to the mount cycle) plus thin `Gsap`/`GsapTimeline` components. Engine-specific vocabulary in both cases: Motion's props are its API and port faithfully, GSAP's API is code and only ever maps lossily onto attributes.
